move the texts to a different file in /data/ and import them accordingly
