+++ components/Text.js
@@ -1,15 +1,5 @@
 import { StyledText } from "./StyledText";
 
-export const texts = {
-  intro:
-    "Give us three Keywords for our AI to SH!T YOUR AI UP. or sth, idunnu, whoami",
-
-  ChooseFourIntro: "Choose your Style!",
-
-  ChooseFourBottom:
-    "Click on the one you like the most to see it in Detail and get more Options.",
-  Variations: "Here are 4 of your Variations",
-};
 export default function Text({ texts }) {
   return <StyledText>{texts}</StyledText>;
 }
