+++ pages/Variations/[imageID].js
@@ -1,4 +1,5 @@
-import Text, { texts } from "@/components/Text";
+import Text from "@/components/Text";
+import { texts } from "../../data/text";
 
 import FourPicsTable from "@/components/FourPicsTable";
 import Header from "@/components/Header";
